dcm4che/dcm4chee-arc-light#341: Failed to store objects with Bulkdata in (0400,0561) Original Attributes Sequence

+++ dcm4che-imageio/src/main/java/org/dcm4che3/imageio/codec/Transcoder.java
@@ -63,6 +63,7 @@ import java.awt.*;
 import java.awt.color.ColorSpace;
 import java.awt.image.*;
 import java.io.*;
+import java.util.List;
 
 /**
  * @author Gunter Zeilinger <gunterze@gmail.com>
@@ -91,6 +92,8 @@ public class Transcoder implements Closeable {
     private boolean closeInputStream = true;
 
     private boolean closeOutputStream = true;
+
+    private boolean deleteBulkDataFiles = true;
 
     private String destTransferSyntax;
 
@@ -189,6 +192,14 @@ public class Transcoder implements Closeable {
         this.closeOutputStream = closeOutputStream;
     }
 
+    public boolean isDeleteBulkDataFiles() {
+        return deleteBulkDataFiles;
+    }
+
+    public void setDeleteBulkDataFiles(boolean deleteBulkDataFiles) {
+        this.deleteBulkDataFiles = deleteBulkDataFiles;
+    }
+
     public boolean isIncludeFileMetaInformation() {
         return includeFileMetaInformation;
     }
@@ -232,6 +243,10 @@ public class Transcoder implements Closeable {
 
     public void setPixelDataBulkDataURI(String pixelDataBulkDataURI) {
         this.pixelDataBulkDataURI = pixelDataBulkDataURI;
+    }
+
+    public List<File> getBulkDataFiles() {
+        return dis.getBulkDataFiles();
     }
 
     private void initDecompressor() {
@@ -294,8 +309,9 @@ public class Transcoder implements Closeable {
             verifier.dispose();
         if (closeInputStream)
             SafeClose.close(dis);
-        for (File tmpFile : dis.getBulkDataFiles())
-            tmpFile.delete();
+        if (deleteBulkDataFiles)
+            for (File tmpFile : dis.getBulkDataFiles())
+                tmpFile.delete();
         if (closeOutputStream && dos != null)
             dos.close();
     }
